Fix pipeline_error_observed 'reconciler' and 'name' tag for multiple sync user
When Config Sync is running in multiple sync mode, the name will be custom name assigned by user, instead of being just `root-sync` or `<namespace-name>`.

This change computes the reconciler name and type based on the namespace and name given. The value of both tags are fixed to be in sync with Config Sync.

b/230385546

Change-Id: I36804f83a0fa5938eddd966dc48e5291dd7be016
GitOrigin-RevId: 9eac9aab514af8cc8c574f2ef179b1bfadf03ccd