test: refactor unit tests to use testing framework
This change refactors the unit tests to consistently use the testing
framework from the standard library rather than ginkgo. A subset of the
tests were already using the testing framework and did not need to be
changed. This change is intended to prepare for migrating the controller
code and tests to Config Sync, which uses the standard testing
framework.

Bug: b/307792024
Change-Id: If1ad002c8cbe47f5e24141686076ae8dd8557953
GitOrigin-RevId: ebed78f153f989f3da404f069e4851556dd8d241